ipa: Create a camera sensor helper class
Setting analogue gain for a specific sensor is not a straghtforward
operation, as one needs to know how the gain is calculated for it.

This commit introduces a new camera sensor helper in libipa which aims
to solve this specific issue.
It is based on the MIPI alliance Specification for Camera Command Set
and implements, for now, only the analogue "Global gain" mode.

Is makes it possible to only have 4 parameters to store per sensor, and
the gain calculation is then done identically for all of them. This
commit gives the gains for imx219 (based on RPi cam_helper), ov5670 and
ov5693.

Adding a new sensor is pretty straightfoward as one only needs to
implement the sub-class for it and register that class to the
CameraSensorHelperFactory.

+++ src/ipa/ipu3/ipu3_agc.h
@@ -32,7 +32,7 @@ public:
 	~IPU3Agc() = default;
 
 	void initialise(struct ipu3_uapi_grid_config &bdsGrid, const IPAConfigInfo &configInfo);
-	void process(const ipu3_uapi_stats_3a *stats, uint32_t &exposure, uint32_t &analogueGain);
+	void process(const ipu3_uapi_stats_3a *stats, uint32_t &exposure, double &analogueGain);
 	bool converged() { return converged_; }
 	bool updateControls() { return updateControls_; }
 	/* \todo Use a metadata exchange between IPAs */
@@ -41,7 +41,7 @@ public:
 private:
 	void processBrightness(const ipu3_uapi_stats_3a *stats);
 	void filterExposure();
-	void lockExposureGain(uint32_t &exposure, uint32_t &gain);
+	void lockExposureGain(uint32_t &exposure, double &gain);
 	void generateStats(const ipu3_uapi_stats_3a *stats);
 	void clearStats();
 	void generateZones(std::vector<RGB> &zones);
